backend/requirements.txt
Django==5.0.2
pymongo==3.12.1
djongo
djangorestframework
pyjwt